requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib.pyplot
sklearn.cluster
sklearn.preprocessing